requirements.txt
python-dotenv
google-api-python-client 
google-auth 
google-auth-oauthlib
google-auth-httplib2 
tqdm>=4.65.0
requests>=2.28.0
matplotlib>=3.5.0 
pandas>=1.4.0
psutil>=5.9.0
speedtest-cli>=2.1.3